+++ tests/test_templates.py
@@ -110,3 +110,22 @@ def test_amplitude_encoding(backend):
         figs, 2, tc.array_to_tensor(np.array([0, 3, 1, 2]), dtype="int32")
     )
     np.testing.assert_allclose(states[0], 1 / np.sqrt(2) * np.array([1, 1, 0, 0]))
+
+
+@pytest.mark.parametrize("backend", [lf("tfb"), lf("jaxb")])
+def test_mpo_measurement(backend):
+    def f(theta):
+        mpo = tc.quantum.QuOperator.from_local_tensor(
+            tc.array_to_tensor(tc.gates._x_matrix), [2, 2, 2], [0]
+        )
+        c = tc.Circuit(3)
+        c.ry(0, theta=theta)
+        c.H(1)
+        c.H(2)
+        e = tc.templates.measurements.mpo_expectation(c, mpo)
+        return e
+
+    v, g = tc.backend.jit(tc.backend.value_and_grad(f))(tc.backend.ones([]))
+
+    np.testing.assert_allclose(v, 0.84147, atol=1e-4)
+    np.testing.assert_allclose(g, 0.54032, atol=1e-4)