+++ tensorcircuit/templates/__init__.py
@@ -1,4 +1,5 @@
 from . import blocks
+from . import chems
 from . import dataset
 from . import graphs
 from . import measurements