+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.0.220311
+
 ### Added
 
 - add sigmoid method on backends